% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rahu_Kalam.R
\name{calculate_rahu_kalam}
\alias{calculate_rahu_kalam}
\title{Calculate Rahu Kalam timing based on sunrise and sunset}
\arguments{
\item{date}{Date for which Rahu Kalam is calculated.}

\item{lat}{Latitude of the location.}

\item{lon}{Longitude of the location.}

\item{tz}{Time zone.}
}
\value{
List with Rahu Kalam start and end times.
}
\description{
Calculate Rahu Kalam timing based on sunrise and sunset
}
